# api/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: |
    Lean Go backend for the core fitness tracking loop: users, exercise library,
    workouts, workout exercises, workout sets, meals, and weight entries.
servers:
  - url: http://localhost:8080
...
paths:
  /v1/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Get Leaderboard
      operationId: getLeaderboard
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [weekly, monthly, yearly, alltime]
            default: weekly
        - name: pillar
          in: query
          required: false
          schema:
            type: string
            enum: [all, training, nutrition, consistency]
            default: all
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user and create a session
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created and session established.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with password and optional 2FA code (session-based)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded. Returns access and refresh tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "202":
          description: Password accepted and 2FA is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an existing session
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access and refresh tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout of current or all sessions
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/sessions:
    get:
      tags: [Auth]
      summary: List active sessions for the user
      operationId: getSessions
      responses:
        "200":
          description: List of active sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSessions"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a specific session
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Start 2FA setup for the authenticated user
      operationId: setupTwoFactor
      responses:
        "201":
          description: Secret and otpauth URL created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorSetupResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Verify 2FA setup with a TOTP code
      operationId: verifyTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCodeRequest"
      responses:
        "200":
          description: 2FA enabled and recovery codes generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorVerifyResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/disable:
    post:
      tags: [Auth]
      summary: Disable 2FA for the authenticated user
      operationId: disableTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCodeRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/recover:
    post:
      tags: [Auth]
      summary: Login with a recovery code
      operationId: recoverTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorRecoveryLoginRequest"
      responses:
        "200":
          description: Login succeeded with recovery code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      parameters:
        - in: query
          name: email
          schema:
            type: string
          description: Exact email filter.
        - in: query
          name: name
          schema:
            type: string
          description: Case-sensitive partial name filter.
      responses:
        "200":
          description: Users in descending creation order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: User detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/records:
    get:
      tags: [Analytics]
      summary: Get user personal records
      operationId: getUserRecords
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: List of personal records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPersonalRecords"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/workout-stats:
    get:
      tags: [Analytics]
      summary: Get user workout stats
      operationId: getUserWorkoutStats
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Workout statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutStats"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/activity-calendar:
    get:
      tags: [Adherence]
      summary: Get user activity calendar
      operationId: getActivityCalendar
      parameters:
        - $ref: "#/components/parameters/UserID"
        - in: query
          name: start
          schema:
            type: string
            format: date
        - in: query
          name: end
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Map of dates to activity types.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/streaks:
    get:
      tags: [Adherence]
      summary: Get user streaks and adherence summary
      operationId: getUserStreaks
      parameters:
        - $ref: "#/components/parameters/UserID"
        - in: query
          name: date
          schema:
            type: string
            format: date
          description: Reference date for streak calculation (defaults to today).
      responses:
        "200":
          description: Streaks and adherence summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreaksResponse"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/chat:
    post:
      tags: [Chat]
      summary: Send a message to the AI coach
      operationId: chat
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/chat/history:
    get:
      tags: [Chat]
      summary: Get chat conversation history
      operationId: getChatHistory
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: query
          schema:
            type: string
            format: uuid
          description: Optional. If provided, returns messages for this conversation. Otherwise returns a list of conversations.
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for the conversation list when `conversation_id` is omitted.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of conversations to return per page when `conversation_id` is omitted.
      responses:
        "200":
          description: List of conversations or conversation detail with messages.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PaginatedConversations"
                  - $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/chat/feedback:
    post:
      tags: [Chat]
      summary: Provide feedback on an AI message
      operationId: chatFeedback
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatFeedbackRequest"
      responses:
        "200":
          description: Feedback received.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/coach-summary:
    get:
      tags: [AI Coach]
      summary: Get AI-generated coach summary for a user
      operationId: getCoachSummary
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Optional. Date for the summary (defaults to today).
      responses:
        "200":
          description: Coach summary data.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/exercises:
    get:
      tags: [Exercises]
      summary: List exercises
      operationId: listExercises
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: muscle_group
          schema:
            type: string
        - in: query
          name: equipment
          schema:
            type: string
        - in: query
          name: difficulty
          schema:
            type: string
      responses:
        "200":
          description: Exercises in ascending name order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedExercises"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Exercises]
      summary: Create exercise
      operationId: createExercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "201":
          description: Created exercise.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "400":
          $ref: "#/components/responses/Error"
...
  /v1/exercises/{id}/history:
    get:
      tags: [Exercises]
      summary: Get exercise history
      operationId: getExerciseHistory
      parameters:
        - $ref: "#/components/parameters/ID"
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Ordered past sessions for the exercise.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedExerciseHistory"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/exercises/program:
    post:
      tags: [Exercises]
      summary: Generate a workout program
      operationId: generateProgram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateProgramRequest"
      responses:
        "200":
          description: Generated workout program.
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
...
    NoContent:
      description: Resource deleted successfully.
  schemas:
...
    AuthSessionResponse:
      type: object
      properties:
        token:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        two_factor_required:
          type: boolean
        two_factor_token:
          type: string
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        avatar:
          type: string
        age:
          type: integer
        date_of_birth:
          type: string
          format: date
        weight:
          type: number
        height:
          type: number
        goal:
          type: string
        activity_level:
          type: string
        tdee:
          type: integer
...
    LeaderboardResponse:
      type: object
      properties:
        period:
          type: string
        pillar:
          type: string
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT